Conflicts:
	bigbluebutton-config/web/index.html

+++ bigbluebutton-apps/src/main/java/org/bigbluebutton/conference/service/presentation/PresentationRoom.java
@@ -1,22 +1,24 @@
-/*
- * BigBlueButton - http://www.bigbluebutton.org
- * 
- * Copyright (c) 2008-2009 by respective authors (see below). All rights reserved.
- * 
- * BigBlueButton is free software; you can redistribute it and/or modify it under the 
- * terms of the GNU Lesser General Public License as published by the Free Software 
- * Foundation; either version 3 of the License, or (at your option) any later 
- * version. 
- * 
- * BigBlueButton is distributed in the hope that it will be useful, but WITHOUT ANY 
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A 
- * PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public License along 
- * with BigBlueButton; if not, If not, see <http://www.gnu.org/licenses/>.
- *
- * $Id: $
- */
+/** 
+* ===License Header===
+*
+* BigBlueButton open source conferencing system - http://www.bigbluebutton.org/
+*
+* Copyright (c) 2010 BigBlueButton Inc. and by respective authors (see below).
+*
+* This program is free software; you can redistribute it and/or modify it under the
+* terms of the GNU Lesser General Public License as published by the Free Software
+* Foundation; either version 2.1 of the License, or (at your option) any later
+* version.
+*
+* BigBlueButton is distributed in the hope that it will be useful, but WITHOUT ANY
+* WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+* PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
+*
+* You should have received a copy of the GNU Lesser General Public License along
+* with BigBlueButton; if not, see <http://www.gnu.org/licenses/>.
+* 
+* ===License Header===
+*/
 
 package org.bigbluebutton.conference.service.presentation;
 
@@ -44,10 +46,10 @@ public class PresentationRoom {
 	int currentSlide = 0;
 	Boolean sharing = false;
 	String currentPresentation = "";
-	long xOffset = 0;
-	long yOffset = 0;
-	long widthRatio = 0;
-	long heightRatio = 0;
+	Double xOffset = 0D;
+	Double yOffset = 0D;
+	Double widthRatio = 0D;
+	Double heightRatio = 0D;
 	ArrayList<String> presentationNames = new ArrayList<String>();
 	
 	public PresentationRoom(String name) {
@@ -94,7 +96,7 @@ public class PresentationRoom {
         }           
     }
 	
-	public void resizeAndMoveSlide(Long xOffset, Long yOffset, Long widthRatio, Long heightRatio) {
+	public void resizeAndMoveSlide(Double xOffset, Double yOffset, Double widthRatio, Double heightRatio) {
 		this.xOffset = xOffset;
 		this.yOffset = yOffset;
 		this.widthRatio = widthRatio;
@@ -191,19 +193,19 @@ public class PresentationRoom {
 		return currentPresenter;
 	}
 
-	public long getxOffset() {
+	public Double getxOffset() {
 		return xOffset;
 	}
 
-	public long getyOffset() {
+	public Double getyOffset() {
 		return yOffset;
 	}
 
-	public long getWidthRatio() {
+	public Double getWidthRatio() {
 		return widthRatio;
 	}
 
-	public long getHeightRatio() {
+	public Double getHeightRatio() {
 		return heightRatio;
 	}
 }
